Improve handling of file redirections
Instead of making the caller do all the work of closing the file
descriptors safely, add `FD_move_safely` mode. This is like Lwt's
`FD_move`, except that it copes with duplicates and records whether the
close happened so it can still be handled later by a finally block if an
exception is thrown.