Introduce Docker backend
This allows to use Docker as a backend for OCluster.

Currently, the store and the sandbox are based on Docker and they must
be used together. The Docker backend is selected at compile-time on
Windows, and the {btrfs,zfs}/runc backends are preferred on other
systems. It is possible to force the use of the Docker backend at
compile-time.

All images, containers, and volumes used by OBuilder are prefixed with
"obuilder=".

The backend has been tested on Windows and Docker.