Docker backend: support copying from context and other build stages
Commit operation need to restore ENTRYPOINT and CMD.

The current implementation uses a temporary volume to store the
tarball of files being copyied. It could be possible to pipe the
tarball into the receiving container instead.

Sadly, Windows bsdtar doesn't support symlinks so we need to carry
around a Docker volume with Cygwin tools.