Attach stress.exe to obuilder package
ocaml-ci executes this if only obuilder-spec can be build:

    opam exec -- dune build --only-packages=obuilder-spec @install @check @runtest && rm -rf _build

Dune would try to build stress.exe as part of the obuider-spec package
@check alias. However, stress links with the obuilder library and such
is part of the obuilder package. The easiest way to express it is to
make it a no-op test of the obuilder package, so that it is not build
with obuilder-spec.